requirements.txt
apache-airflow~=2.10.2
requests~=2.32.3